api_server/requirements.txt
sanic==23.3.0
sanic-ext==23.6.0
sanic_cors
googletrans==3.1.0a0
fake_useragent
elasticsearch
python-dotenv